requirements.txt
# ============================================
# Potato Disease Detection - Requirements
# ============================================

# Deep Learning & ML
tensorflow>=2.10.0,<2.16.0
numpy>=1.23.0,<2.0.0
matplotlib>=3.6.0

# Backend API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image Processing
Pillow>=10.0.0

# Google Gemini AI
google-generativeai>=0.3.0

# Environment Variables
python-dotenv>=1.0.0

# Translation Services
deep-translator>=1.11.0

# Data Validation
pydantic>=2.0.0
